Optimized for win10 and performance.
The previous code did not work on win10, probably due to enhanced security. So I removed the WMI code to toggle the network interface ON/OFF and replaced it with a NetShell command. I also noticed my code in the previous version did not really made sense in some parts so I changed that as well. I also removed the Timer subroutine and replaced that with WinWaitActive and Process,WaitClose commands for better performance.