Best way to send singe or double click
Added my own method for sending key's using mouse clicks. Best way to do this is by using keywait and errorlevel.

Most of the other examples have some form of delay witch makes them slow. The code at the top
is my own where I make use of Keywait and the fact that when ErrorLevel is set to zerro the code below is
executed right away. So there is zero to none waiting time for actions to be trigged.