Changed how script starts
When the script loads for the first time or after the setting.ini is removed. The script will run normally so exe files can be dropped on it. Once a game.exe is added to the script.  The script will automaticity starts with admin privileges. This is to ensure the game window can accepts input. And 9 out of 10 games need admin anyway for AHK to work properly. So enabling it by default.